fix: compare the total the label displays, not its nearest double (PCL018) (#29)
* a pattern must not match digits the code cannot read

Python's Unicode digit class escape matches 680 characters, of which ten are
ASCII, and float and Decimal convert all 680. Everything downstream of these
patterns reads ASCII only: _phones strips [^0-9] from what _PHONE matched,
and a matched percentage is converted before it is compared.

The pattern was therefore wider than the code deriving a value from it, and
PCL002 read that as a pass. A telephone number written in another numeral
system matched _PHONE, the ASCII-only strip reduced it to the empty string,
and the empty string counted as one of the two distinct numbers section
1393.1(c)(4) lists:

    _phones -> ['', '5555550100']
    PCL002  -> conforms: 2 distinct telephone numbers appear on the label.

One number had been read. This is the same class of defect as the email
address that counted as a website: a check reporting ok on the document it
exists to notice.

_PERCENT, _PHONE and _YEAR_TITLE now write [0-9]. Narrowed, a figure the
tool cannot read is reported absent or not evaluated rather than silently
valued: a total row in another numeral system makes PCL018 not evaluated,
which is the direction ADR 0001 requires. PCL017 is the one check where the
narrowing produces a deviation rather than a not-evaluated, and that is
defensible on its own cited basis, section 1393.1(i), that the format of the
issued label may not be altered.

The rule is enforced rather than remembered: tests/test_repo_hygiene.py
scans src and scripts, because this failure is invisible on every input
anyone tries. The test fixtures build their non-ASCII digits from the code
point so the test file stays ASCII and the linter's ambiguous-character rule
is not silenced to let a test in.

Six new tests fail against the pre-change source and pass against this one.
Four positive controls, the ASCII case of each behavioural test, pass in
both states. Conclusions about all ten cached published labels are
unchanged.


* compare the total the label displays, not its nearest double

_PERCENT places no limit on the digits after the point, and a double cannot
hold twenty of them. float("99.999999999999999999") == 100.0 is True, so
PCL018 reported a pass on a displayed total that is not one hundred.

The figures are read as decimals now. Decimal compares by value rather than
by representation, so 100, 100.0 and 100.00 all still conform, and a total
that only rounds to one hundred is reported with the figure the label
displays rather than with the double it collapsed to.

There is deliberately no exception handling around the conversion, and the
code says so. _PERCENT matches [0-9]{1,3} with an optional fractional part,
so what reaches Decimal is always a valid decimal literal and
InvalidOperation cannot be raised. A catch would be unreachable, and if it
ever did run it would drop a total row's figure and let the check conclude
from the ones that remained, which is the opposite of what this file
promises.

Adversarial rather than observed: every published label read for this
project displays whole percentages, which is why issue 21 filed this as low
priority rather than urgent. It is still a check that could report ok on a
document that deviates.

Three new tests fail against the pre-change source and pass against this
one. Eleven pass in both, including four renderings of one hundred and a
whole-number deviation, which are the controls on Decimal comparing by value
and on the finding still naming what it saw. Conclusions about all ten
cached published labels are unchanged.

Fixes #21


---------

+++ src/power_content_check/checks.py
@@ -32,6 +32,7 @@ from __future__ import annotations
 import re
 from collections.abc import Callable
 from dataclasses import dataclass
+from decimal import Decimal
 
 from .citations import issued_format, reg
 from .extract import LabelDocument
@@ -609,16 +610,31 @@ def _pcl018(doc: LabelDocument, ctx: CheckContext) -> CheckResult:
     # test_multiple_conforming_total_rows_conforms): a repeated line is not
     # assumed to be a duplicate-extraction artifact of the same row, since a
     # label can genuinely carry more than one identically-formatted total.
-    rows: list[tuple[str, list[float]]] = []
+    #
+    # The figures are read as decimals, not as doubles. _PERCENT places no
+    # limit on the digits after the point, and a double cannot hold twenty of
+    # them: float("99.999999999999999999") == 100.0 is True, so a displayed
+    # total that is not one hundred would be reported as a pass. Decimal
+    # compares the figure the label displays, and by value rather than by
+    # representation, so 100, 100.0 and 100.00 all still conform.
+    #
+    # There is deliberately no exception handling around the conversion.
+    # _PERCENT matches [0-9]{1,3} with an optional fractional part, so what
+    # reaches Decimal is always a valid decimal literal and InvalidOperation
+    # cannot be raised here. Catching it would be unreachable code that, if it
+    # ever did run, would drop a total row's figure and let the check conclude
+    # from the ones that remained, which is the opposite of what this file
+    # promises.
+    rows: list[tuple[str, list[Decimal]]] = []
     for line in doc.normalized_lines:
         match = _TOTAL_ROW.match(_row_label(line))
         if not match:
             continue
         row_text = match.group(0)
-        values = [float(v.rstrip("% ").strip()) for v in re.findall(_PERCENT, match.group(1))]
+        values = [Decimal(v.rstrip("% ").strip()) for v in re.findall(_PERCENT, match.group(1))]
         rows.append((row_text, values))
     if rows:
-        deviations = [(row, v) for row, values in rows for v in values if v != 100.0]
+        deviations = [(row, v) for row, values in rows for v in values if v != 100]
         if deviations:
             cited = "; ".join(f"{row!r} has {v}" for row, v in deviations)
             return _bad(
